requirements.txt
absl-py==1.4.0
astor==0.8.1
astunparse==1.6.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cycler==0.11.0
deap==1.4.1
Django==2.1.7
exceptiongroup==1.1.3
filelock==3.12.2
Flask==2.2.5
flatbuffers==2.0.7
fonttools==4.38.0
fsspec==2023.1.0
future==0.18.3
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpt-2-simple==0.8.1
grpcio==1.57.0
h5py==2.10.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.2
importlib-metadata==6.7.0
imutils==0.5.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.4
joblib==1.3.2
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
nltk==3.8.1
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
playsound==1.3.0
pluggy==1.2.0
protobuf==3.16.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyMySQL==0.9.3
pyparsing==3.1.1
pyswarms==1.1.0
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-learn==0.22.2.post1
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post7
sklearn-extensions==0.0.2
sklearn-genetic==0.2
SwarmPackagePy==1.0.0a5
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0rc1
tensorflow-estimator==2.7.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
toposort==1.10
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
transformers==4.30.2
typing_extensions==4.7.1
ultralytics==8.0.137
urllib3==2.0.6
Werkzeug==2.2.3
wrapt==1.15.0
xgboost==1.6.2
zipp==3.15.0
